+++ app/src/main/java/com/shootr/mobile/ui/fragments/StreamTimelineFragment.java
@@ -8,8 +8,10 @@ import android.graphics.Color;
 import android.graphics.Typeface;
 import android.net.Uri;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.widget.SwipeRefreshLayout;
+import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.RecyclerView;
 import android.text.Editable;
 import android.text.TextWatcher;
@@ -256,6 +258,18 @@ public class StreamTimelineFragment extends BaseFragment
         return false;
       }
     });
+    shotsTimeline.setItemAnimator(new DefaultItemAnimator() {
+      @Override
+      public boolean canReuseUpdatedViewHolder(@NonNull RecyclerView.ViewHolder viewHolder) {
+        return true;
+      }
+
+      @Override
+      public boolean canReuseUpdatedViewHolder(@NonNull RecyclerView.ViewHolder viewHolder,
+          @NonNull List<Object> payloads) {
+        return true;
+      }
+    });
     newShotBarContainer.setVisibility(View.GONE);
     return fragmentView;
   }
@@ -280,6 +294,9 @@ public class StreamTimelineFragment extends BaseFragment
     Integer streamMode = getArguments().getInt(EXTRA_READ_WRITE_MODE, 0);
     setStreamTitleClickListener(idStream);
     setupPresentersInitialization(idStream, streamAuthorIdUser, streamMode);
+  }
+
+  @Override public void sendAnalythicsEnterTimeline() {
     sendTimelineAnalytics();
   }
 
@@ -544,7 +561,7 @@ public class StreamTimelineFragment extends BaseFragment
         }, //
         new OnNiceShotListener() {
           @Override public void markNice(ShotModel shot) {
-            streamTimelinePresenter.markNiceShot(shot.getIdShot());
+            streamTimelinePresenter.markNiceShot(shot);
             sendNiceAnalytics(shot);
           }
 
@@ -1099,6 +1116,14 @@ public class StreamTimelineFragment extends BaseFragment
       CustomActionItemBadge.updateFilterAlert(getActivity(), showHoldingShotsMenuItem,
           showHoldingShotsMenuItem.getIcon(), " ");
     }
+  }
+
+  @Override public void renderNice(ShotModel shotModel) {
+    adapter.markNice(shotModel);
+  }
+
+  @Override public void renderUnnice(String idShot) {
+    adapter.unmarkNice(idShot);
   }
 
   private void hideFilterAlert() {